adk/contract_qna/requirements.txt
google-cloud-storage
reportlab
google-cloud-aiplatform
google-adk